fixed death message spam

+++ src/liblogic.cpp
@@ -95,7 +95,7 @@ void liblogic_init(shared_ptr<gamestate> g) {
 
     //#ifdef SPAWN_MONSTERS
 
-    for (int i = 0; i < g->dungeon->floors->size(); i++) {
+    for (int i = 0; i < (int)g->dungeon->floors->size(); i++) {
         create_npc_at_with(g, RACE_ORC, (vec3){10, 10, i}, orc_init_test);
         create_npc_at_with(g, RACE_ORC, (vec3){10, 9, i}, orc_init_test);
         create_npc_at_with(g, RACE_ORC, (vec3){11, 9, i}, orc_init_test);
